server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
charset-normalizer==3.4.1
cleo==2.1.0
click==8.3.1
colorama==0.4.6
comm==0.2.3
crashtest==0.4.1
debugpy==1.8.17
decorator==5.2.1
distlib==0.3.9
dulwich==0.21.7
executing==2.2.1
fastjsonschema==2.21.1
filelock==3.17.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.1
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
keyring==24.3.1
matplotlib-inline==0.2.1
more-itertools==10.6.0
msgpack==1.1.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.5
pexpect==4.9.0
pkginfo==1.12.1.2
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
pyzmq==27.1.0
RapidFuzz==3.12.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
tomlkit==0.13.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.3.13.13
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.40.0
virtualenv==20.29.3
wcwidth==0.2.14
zipp==3.21.0
google-generativeai
google-genai
fastapi==0.128.0
python-dotenv
python-decouple
asyncpg

openai
PyMuPDF
pytesseract
Pillow
langchain-text-splitters
pgvector
python-multipart
psycopg2-binary
pandas
openpyxl
lxml
aiohttp
pymilvus>=2.4.0
python-hwpx>=2.9.0
python-pptx>=1.0.0
